chore: reorganize EventListener files
There were 2 `CoroutineExtensions.kt`... Now it's renamed.
